+++ CHANGES.rst
@@ -9,6 +9,27 @@ Changes
 
 * Nothing yet.
 
+`2.1.0 <https://github.com/Ouranosinc/Magpie/tree/2.1.0>`_ (2020-09-22)
+------------------------------------------------------------------------------------
+
+Features / Changes
+~~~~~~~~~~~~~~~~~~~~~
+
+* Adjust ``alembic`` migration scripts to employ date-ordered naming convention to help searching features within them.
+* Add ``DENY`` permission-access concept with new ``PermissionSet`` object, ``Access`` and ``Scope`` enums.
+* Remove ``-match`` suffixed entries from ``Permission`` enum in favor of new ``PermissionSet`` definition.
+* Update permission entries to employ explicit representation as ``[name]-[access]-[scope]`` in the database
+  (`#342 <https://github.com/Ouranosinc/Magpie/issues/342>`_).
+* Provide new ``permissions`` list in applicable API responses, with explicit ``name``, ``access`` and ``scope`` fields
+  for each ``PermissionSet``. Responses will also return the concatenated *explicit* string representations (see above)
+  on top of older *implicit*  representation still returned in ``permission_names`` field for backward compatibility
+  (note: ``DENY`` elements will only be represented as *explicit* as there was no *implicit* permissions of this type).
+* | Upgrade migration script is added to convert the implicit names to new explicit permission names.
+  |
+  | **WARNING**:
+  | Downgrade migration drops any ``DENY`` permission that would be added in future versions,
+    as they do not exist prior to this introduced version.
+
 `2.0.1 <https://github.com/Ouranosinc/Magpie/tree/2.0.1>`_ (2020-09-30)
 ------------------------------------------------------------------------------------